backend/pyproject.toml
[tool.poetry]
name = "parity-test-task"
version = "0.1.0"
description = ""
authors = ["zewa-bamuga <tikhonov.igor2028@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
django-cors-headers = "^4.6.0"
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
